third_party/WebKit/LayoutTests/virtual/threaded/fast/scroll-behavior/smooth-scroll/mousewheel-scroll-interrupted-expected.txt

This test ensures that input driven user scrolls get cancelled by non-input driven scrolls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getScrollAnimationState(document) == RUNNING_ON_COMPOSITOR is true
PASS isOrWillBeCancelled(document) is true
PASS document.scrollingElement.scrollTop < 40 && document.scrollingElement.scrollLeft == 0 became true
PASS successfullyParsed is true

TEST COMPLETE
